@@ -1,27 +1,26 @@
 import 'package:flutter/material.dart';
+import 'package:rubin_chart/rubin_chart.dart';
 
-class ChartTheme {
+class AppTheme {
   final ThemeData themeData;
   final TextStyle _queryStyle;
-  final TextStyle _editorTitleStyle;
   final TextStyle _titleStyle;
+  final TextStyle _axisLabelStyle;
   final double querySpacerWidth;
   final Duration animationSpeed;
   final Offset newWindowOffset;
   final Size newPlotSize;
   final double resizeInteractionWidth;
+  final double toolbarHeight;
+  final Color wireColor;
+  final double wireThickness;
+  // Chart settings
+  final ChartTheme chartTheme;
 
-  const ChartTheme({
+  const AppTheme({
     required this.themeData,
     TextStyle queryStyle = const TextStyle(
       fontSize: 20,
-      fontStyle: FontStyle.normal,
-      decoration: TextDecoration.none,
-      inherit: false,
-      height: 1.0,
-    ),
-    TextStyle editorTitleStyle = const TextStyle(
-      fontSize: 30,
       fontStyle: FontStyle.normal,
       decoration: TextDecoration.none,
       inherit: false,
@@ -34,26 +33,30 @@ class ChartTheme {
       inherit: false,
       height: 1.0,
     ),
+    TextStyle axisLabelStyle = const TextStyle(
+      fontSize: 20,
+      fontStyle: FontStyle.normal,
+      decoration: TextDecoration.none,
+      inherit: false,
+      height: 1.0,
+    ),
     this.querySpacerWidth = 10,
     this.animationSpeed = const Duration(milliseconds: 500),
     this.newWindowOffset = const Offset(20, 20),
     this.newPlotSize = const Size(600, 400),
     this.resizeInteractionWidth = kMinInteractiveDimension / 4,
+    this.toolbarHeight = 40,
+    this.wireColor = Colors.red,
+    this.wireThickness = 4,
+    this.chartTheme = const ChartTheme(),
   })  : _queryStyle = queryStyle,
-        _editorTitleStyle = editorTitleStyle,
-        _titleStyle = titleStyle;
+        _titleStyle = titleStyle,
+        _axisLabelStyle = axisLabelStyle;
 
-  TextStyle get queryStyle =>
-      _queryStyle.copyWith(color: themeData.colorScheme.primary);
-  TextStyle get queryOperatorStyle =>
-      _queryStyle.copyWith(color: themeData.colorScheme.secondary);
-  TextStyle get editorTitleStyle =>
-      _editorTitleStyle.copyWith(color: themeData.colorScheme.primary);
-  TextStyle get titleStyle =>
-      _titleStyle.copyWith(color: themeData.colorScheme.primary);
-
-  Color get selectionColor => themeData.colorScheme.primaryContainer;
-  Color get selectionEdgeColor => themeData.primaryColor;
+  TextStyle get queryStyle => _queryStyle.copyWith(color: themeData.colorScheme.primary);
+  TextStyle get queryOperatorStyle => _queryStyle.copyWith(color: themeData.colorScheme.secondary);
+  TextStyle get titleStyle => _titleStyle.copyWith(color: themeData.colorScheme.primary);
+  TextStyle get axisLabelStyle => _axisLabelStyle.copyWith(color: themeData.colorScheme.primary);
 
   InputDecorationTheme get queryTextDecorationTheme => InputDecorationTheme(
         border: OutlineInputBorder(
